services/payment/payment/payment/serializers.py
from rest_framework import serializers

from payment.payment.models import Payment


class PaymentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Payment
        exclude = ('is_deleted',)
