services/payment/requirements/base.txt
asgiref==3.5.2
Django==3.0.10
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-environ==0.8.1
django-model-utils==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.1.1
psycopg2==2.8.6
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.1.0

